@@ -1,6 +1,6 @@
 {
   "name": "annotated-meals",
-  "version": "0.10.1",
+  "version": "1.0.0",
   "description": "write down recipes while watching youtube cooking videos",
   "main": "index.js",
   "scripts": {
@@ -8,7 +8,6 @@
     "build": "npm run webpack:build & npm run css",
     "css": "npm run sass && npm run autoprefix",
     "sass": "node-sass scss/** -o public/static/css/**",
-    "test": "mocha tests --compilers js:babel-register",
     "webpack:dev": "webpack",
     "webpack:build": "webpack"
   },
@@ -27,23 +26,22 @@
   },
   "homepage": "https://github.com/psherman/annotated-meals#readme",
   "devDependencies": {
-    "autoprefixer": "^6.0.3",
-    "babel-core": "^6.13.2",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-react": "^6.11.1",
-    "babel-register": "^6.11.6",
+    "autoprefixer": "^6.6.1",
+    "babel-core": "^6.21.0",
+    "babel-loader": "^6.2.10",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
     "chai": "^3.4.0",
     "mocha": "^2.3.3",
-    "node-sass": "^3.3.3",
-    "postcss-cli": "^2.1.0",
+    "node-sass": "^4.1.1",
+    "postcss-cli": "^2.6.0",
     "sinon": "^1.17.2",
-    "webpack": "^1.12.2"
+    "webpack": "^1.14.0"
   },
   "dependencies": {
+    "mobx": "^2.7.0",
+    "mobx-react": "^4.1.0",
     "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-redux": "^3.1.2",
-    "redux": "^3.5.2"
+    "react-dom": "^15.3.1"
   }
 }
