@@ -1,1 +1,1 @@
-webpackJsonp([0],{0:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}var l=n(1),r=a(l),u=n(30),i=a(u),s=n(163),c=n(177),o=n(187),d=a(o),f=n(200),p=a(f),m=n(201),v=n(197),h=v.SetupStorage(),E=s.applyMiddleware(m.StorageAPI)(s.createStore)(p["default"],h);i["default"].render(r["default"].createElement(c.Provider,{store:E},r["default"].createElement(d["default"],null)),document.getElementById("content"))},187:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}function l(e){return u["default"].createElement("div",{className:"app"},u["default"].createElement(f["default"],null),u["default"].createElement("div",null,u["default"].createElement(o["default"],null),u["default"].createElement(s["default"],null)))}Object.defineProperty(t,"__esModule",{value:!0}),t["default"]=l;var r=n(1),u=a(r),i=(n(177),n(188)),s=a(i),c=n(196),o=a(c),d=n(198),f=a(d)},188:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}Object.defineProperty(t,"__esModule",{value:!0});var l=n(1),r=a(l),u=n(177),i=n(189),s=a(i),c=n(192),o=a(c),d=n(195),f=a(d),p=r["default"].createClass({displayName:"Annotater",render:function(){var e=this.props.recipe;if(null===e)return null;var t=e.name,n=void 0===t?"":t,a=e.ytID,l=void 0===a?"":a,u=e.ingredients,i=void 0===u?[]:u,c=e.instructions,d=void 0===c?[]:c;return r["default"].createElement("div",{className:"annotater"},r["default"].createElement("div",{className:"edit-view"},r["default"].createElement(f["default"],{ytID:l}),r["default"].createElement(s["default"],{name:n,ingredients:i,instructions:d})),r["default"].createElement("div",{className:"print-view"},r["default"].createElement(o["default"],{name:n,ytID:l,ingredients:i,instructions:d})))}});t["default"]=u.connect(function(e){var t=e.recipes,n=e.index;return{recipe:null!==n?t[n]:null}})(p)},189:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}function l(e){return e.split("\n").filter(function(e){return""!==e})}Object.defineProperty(t,"__esModule",{value:!0});var r=n(1),u=a(r),i=n(177),s=n(190),c=u["default"].createClass({displayName:"LiveEditor",render:function(){var e=this.props,t=e.name,n=void 0===t?"":t,a=e.ingredients,l=void 0===a?[]:a,r=e.instructions,i=void 0===r?[]:r,s=e.setName,c=e.setIngredients,f=e.setInstructions;return u["default"].createElement("div",{className:"live-editor"},u["default"].createElement(o,{name:"name",submit:function(e){s(e)},value:n}),u["default"].createElement(d,{name:"ingredients",submit:function(e){c(e)},value:l.join("\n")}),u["default"].createElement(d,{name:"instructions",submit:function(e){f(e)},value:i.join("\n")}))}});t["default"]=i.connect(null,{setName:s.setName,setIngredients:s.setIngredients,setInstructions:s.setInstructions})(c);var o=u["default"].createClass({displayName:"UserInput",propTypes:{value:u["default"].PropTypes.string.isRequired,name:u["default"].PropTypes.string.isRequired,submit:u["default"].PropTypes.func.isRequired},getInitialState:function(){return{value:this.props.value||""}},componentWillReceiveProps:function(e){this.setState({value:e.value})},handleChange:function(e){this.setState({value:e.target.value})},handleBlur:function(e){this.props.submit(this.state.value)},handleSubmit:function(e){13===e.which&&this.props.submit(this.state.value)},render:function(){return u["default"].createElement("div",null,u["default"].createElement("label",null,this.props.name,u["default"].createElement("input",{type:"text",value:this.state.value,onChange:this.handleChange,onBlur:this.handleBlur,onKeyDown:this.handleSubmit})))}}),d=u["default"].createClass({displayName:"UserTextarea",getInitialState:function(){return{value:this.props.value||""}},componentWillReceiveProps:function(e){this.setState({value:e.value})},handleChange:function(e){this.setState({value:e.target.value})},handleBlur:function(e){this.props.submit(l(this.state.value))},handleSubmit:function(e){13===e.which&&this.props.submit(l(this.state.value))},render:function(){return u["default"].createElement("div",null,u["default"].createElement("label",null,this.props.name,u["default"].createElement("textarea",{value:this.state.value,onChange:this.handleChange,onBlur:this.handleBlur,onKeyDown:this.handleSubmit})))}})},190:function(e,t,n){"use strict";function a(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&(t[n]=e[n]);return t["default"]=e,t}function l(e){return{type:d.CREATE_RECIPE,ytID:e}}function r(e){return{type:d.SET_NAME,name:e}}function u(e){return{type:d.SET_INGREDIENTS,ingredients:e}}function i(e){return{type:d.SET_INSTRUCTIONS,instructions:e}}function s(e){return{type:d.DELETE_RECIPE,index:e}}function c(e){return{type:d.LOAD_RECIPE,index:e}}Object.defineProperty(t,"__esModule",{value:!0}),t.createRecipe=l,t.setName=r,t.setIngredients=u,t.setInstructions=i,t.deleteRecipe=s,t.loadRecipe=c;var o=n(191),d=a(o)},191:function(e,t){"use strict";Object.defineProperty(t,"__esModule",{value:!0});t.CREATE_RECIPE="CREATE_RECIPE",t.SET_NAME="SET_NAME",t.SET_INGREDIENTS="SET_INGREDIENTS",t.SET_INSTRUCTIONS="SET_INSTRUCTIONS",t.DELETE_RECIPE="DELETE_RECIPE",t.LOAD_RECIPE="LOAD_RECIPE"},192:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}function l(e){var t=e.name,n=void 0===t?"":t,a=e.ytID,l=void 0===a?"":a,r=e.ingredients,i=void 0===r?[]:r,c=e.instructions,d=void 0===c?[]:c,f=""!==l?"https://youtu.be/"+l:null;return u["default"].createElement("div",{className:"recipe"},u["default"].createElement("h2",null,n),u["default"].createElement("h3",null,f),u["default"].createElement(s["default"],{values:i}),u["default"].createElement(o["default"],{values:d}))}Object.defineProperty(t,"__esModule",{value:!0}),t["default"]=l;var r=n(1),u=a(r),i=n(193),s=a(i),c=n(194),o=a(c)},193:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}function l(e){return u["default"].createElement("div",{className:"ingredients"},u["default"].createElement("h3",null,"Ingredients"),u["default"].createElement("ul",null,e.values.map(function(e,t){return u["default"].createElement("li",{key:t},e)})))}Object.defineProperty(t,"__esModule",{value:!0}),t["default"]=l;var r=n(1),u=a(r)},194:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}function l(e){return u["default"].createElement("div",{className:"instructions"},u["default"].createElement("h3",null,"Instructions"),u["default"].createElement("ol",null,e.values.map(function(e,t){return u["default"].createElement("li",{key:t},e)})))}Object.defineProperty(t,"__esModule",{value:!0}),t["default"]=l;var r=n(1),u=a(r)},195:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}Object.defineProperty(t,"__esModule",{value:!0});var l=n(1),r=a(l);t["default"]=r["default"].createClass({displayName:"video",propTypes:{ytID:r["default"].PropTypes.string.isRequired},shouldComponentUpdate:function(e,t){return e.ytID!==this.props.ytID},render:function(){var e="https://www.youtube.com/embed/"+this.props.ytID,t=""===this.props.ytID?null:r["default"].createElement("iframe",{width:"560",height:"315",src:e,frameBorder:"0"});return r["default"].createElement("div",{className:"yt"},t)}})},196:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}Object.defineProperty(t,"__esModule",{value:!0});var l=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var a in n)Object.prototype.hasOwnProperty.call(n,a)&&(e[a]=n[a])}return e},r=n(1),u=a(r),i=n(177),s=n(197),c=n(190),o=u["default"].createClass({displayName:"RecipeMenu",propTypes:{recipes:u["default"].PropTypes.array.isRequired,index:u["default"].PropTypes.number},render:function(){var e=this,t=this.props.recipes.map(function(t,n){return null===t?null:u["default"].createElement(d,l({key:n,index:n,active:n===e.props.index,"delete":e.props.deleteRecipe,load:e.props.loadRecipe},t))});return u["default"].createElement("div",{className:"recipe-menu"},"Saved Recipes:",u["default"].createElement("ul",{className:"saved-recipes"},t),u["default"].createElement(f,{createRecipe:this.props.createRecipe}))}});t["default"]=i.connect(function(e){return{recipes:e.recipes,index:e.index}},{deleteRecipe:c.deleteRecipe,loadRecipe:c.loadRecipe,createRecipe:c.createRecipe})(o);var d=u["default"].createClass({displayName:"Thumbnail",shouldComponentUpdate:function(e,t){return this.props.ytID!==e.ytID||this.props.name!==e.name||this.props.index!==e.index||this.props.active!==e.active},handleDelete:function(e){e.stopPropagation(),this.props["delete"](this.props.index)},handleClick:function(e){this.props.load(this.props.index)},render:function(){var e=this.props,t=e.ytID,n=e.name,a=e.active,l="https://i.ytimg.com/vi/"+t+"/mqdefault.jpg",r=""===t?u["default"].createElement("div",{className:"empty-thumb"},"?"):u["default"].createElement("img",{src:l}),i="thumbnail";return a&&(i+=" active"),u["default"].createElement("li",{className:i,onClick:this.handleClick},u["default"].createElement("div",null,r),u["default"].createElement("div",{className:"thumb-info"},n),u["default"].createElement("div",{className:"thumb-controls"},u["default"].createElement("button",{title:"delete recipe",onClick:this.handleDelete},String.fromCharCode(215))))}}),f=u["default"].createClass({displayName:"RecipeCreator",getInitialState:function(){return{value:""}},createRecipe:function(e){e.preventDefault();var t=s.VideoID(this.state.value);this.props.createRecipe(t),this.setState({value:""})},handleChange:function(e){this.setState({value:e.target.value})},render:function(){return u["default"].createElement("div",null,u["default"].createElement("input",{placeholder:"YouTube URL...",value:this.state.value,onChange:this.handleChange}),u["default"].createElement("button",{onClick:this.createRecipe},"Add Recipe"))}})},197:function(e,t){"use strict";function n(){return JSON.parse(localStorage.getItem("recipes"))}function a(e){localStorage.setItem("recipes",JSON.stringify(e))}function l(){var e=n();return null===e&&(e=[],a(e)),{recipes:e,index:null}}function r(e){var t="",n=document.createElement("a");switch(n.href=e,n.hostname){case"www.youtube.com":if(""===n.search)break;var a=n.search.slice(1).split("&");a.some(function(e){var n=e.split("=");return"v"===n[0]?(t=n[1],!0):!1});break;case"youtu.be":var a=n.pathname.split("/");t=a[a.length-1]}return t}Object.defineProperty(t,"__esModule",{value:!0}),t.LoadRecipes=n,t.StoreRecipes=a,t.SetupStorage=l,t.VideoID=r},198:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}Object.defineProperty(t,"__esModule",{value:!0}),t["default"]=function(e){return r["default"].createElement("header",null,r["default"].createElement("h1",null,"Annotated Meals"),r["default"].createElement("p",null,"Did you see a recipe on YouTube that you'd like to make, but don't want to constantly pause the video while you're cooking? Annotated Meals lets you quickly type up a recipe from a video."),r["default"].createElement(i["default"],null),r["default"].createElement("p",null,"For a quick test, try pasting this link ",r["default"].createElement("strong",null,"https://www.youtube.com/watch?v=bjmYkPkjnVo"),' into the "YouTube URL..." input below.'))};var l=n(1),r=a(l),u=n(199),i=a(u)},199:function(e,t,n){"use strict";function a(e){return e&&e.__esModule?e:{"default":e}}Object.defineProperty(t,"__esModule",{value:!0});var l=n(1),r=a(l);t["default"]=r["default"].createClass({displayName:"howto",getInitialState:function(){return{visible:!0}},shouldComponentUpdate:function(e,t){return t.visible!==this.state.visible},toggleVisible:function(e){e.preventDefault(),this.setState({visible:!this.state.visible})},render:function(){var e=this.state.visible?"Hide Help":"Show Help",t=this.state.visible?"":"hidden";return r["default"].createElement("div",{className:"help"},r["default"].createElement("h3",null,"How to:"),r["default"].createElement("button",{onClick:this.toggleVisible},e),r["default"].createElement("ol",{className:t},r["default"].createElement("li",null,"Go to the YouTube page for the recipe you would like to make and copy the url from the address bar."),r["default"].createElement("li",null,'Paste the url into the "YouTube URL..." box below, then click "Add Recipe"'),r["default"].createElement("li",null,"An embedded copy of the video will appear in the page."),r["default"].createElement("li",null,"Type in the name of the recipe."),r["default"].createElement("li",null,"Press play on the video and write down the ingredients and instructions while you watch the video, pausing to catch up whenever the chef is moving faster than you can type."),r["default"].createElement("li",null,'When you\'ve written everything down, click the "Save" button.'),r["default"].createElement("li",null,"If you would like a paper copy, print the recipe. Only your input will be printed, everything else will be hidden.")))}})},200:function(e,t,n){"use strict";function a(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&(t[n]=e[n]);return t["default"]=e,t}function l(){var e=arguments.length<=0||void 0===arguments[0]?i:arguments[0],t=arguments[1],n=e.index,a=void 0===n?null:n,l=e.recipes,r=void 0===l?[]:l;switch(t.type){case u.CREATE_RECIPE:var s=r.length;return Object.assign({},{index:s,recipes:r.concat({name:"",ytID:t.ytID,ingredients:[],instructions:[]})});case u.LOAD_RECIPE:var c=null;return t.index>=0&&t.index<r.length&&(c=t.index),Object.assign({},e,{index:c});case u.DELETE_RECIPE:return Object.assign({},e,{recipes:r.map(function(e,n){return n!==t.index?e:null}),index:t.index===a?null:a});case u.SET_NAME:return Object.assign({},e,{recipes:r.map(function(e,n){return n===a?Object.assign({},e,{name:t.name}):e})});case u.SET_INGREDIENTS:return Object.assign({},e,{recipes:r.map(function(e,n){return n===a?Object.assign({},e,{ingredients:t.ingredients}):e})});case u.SET_INSTRUCTIONS:return Object.assign({},e,{recipes:r.map(function(e,n){return n===a?Object.assign({},e,{instructions:t.instructions}):e})});default:return e}}Object.defineProperty(t,"__esModule",{value:!0}),t["default"]=l;var r=n(191),u=a(r),i={recipes:[],index:null}},201:function(e,t,n){"use strict";function a(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&(t[n]=e[n]);return t["default"]=e,t}Object.defineProperty(t,"__esModule",{value:!0}),t.StorageAPI=void 0;var l=n(191),r=a(l),u=n(197);t.StorageAPI=function(e){return function(t){return function(n){switch(n.type){case r.SET_NAME:case r.SET_INGREDIENTS:case r.SET_INSTRUCTIONS:case r.DELETE_RECIPE:var a=t(n),l=e.getState(),i=l.recipes;return u.StoreRecipes(i.filter(function(e){return null!==e})),a;default:return t(n)}}}}}});
+webpackJsonp([0],{0:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}var a=n(1),r=l(a),u=n(30),i=l(u),s=n(163),c=(n(164),n(165)),o=l(c),d=n(173),f=(0,d.load)("recipes");null===f&&(f=[]);var p=(0,s.observable)({recipes:f,index:null});(0,s.autorun)(function(){(0,d.save)("recipes",p.recipes.filter(function(e){return null!==e}))}),i.default.render(r.default.createElement(o.default,{recipes:p}),document.getElementById("content"))},165:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=n(1),r=l(a),u=n(164),i=n(166),s=l(i),c=n(172),o=l(c),d=n(174),f=l(d);t.default=(0,u.observer)(function(e){var t=e.recipes,n=t.recipes,l=t.index,a=null!==l?n[l]:null;return r.default.createElement("div",{className:"app"},r.default.createElement(f.default,null),r.default.createElement("div",null,r.default.createElement(o.default,{recipes:e.recipes}),r.default.createElement(s.default,{recipe:a})))})},166:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=n(1),r=l(a),u=n(164),i=n(167),s=l(i),c=n(168),o=l(c),d=n(171),f=l(d),p=function(e){var t=e.recipe;return t?r.default.createElement("div",{className:"annotater"},r.default.createElement("div",{className:"edit-view"},r.default.createElement(f.default,{ytID:e.recipe.ytID}),r.default.createElement(s.default,{recipe:e.recipe})),r.default.createElement("div",{className:"print-view"},r.default.createElement(o.default,{recipe:e.recipe}))):null};t.default=(0,u.observer)(p)},167:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=n(1),r=l(a),u=n(164),i=r.default.createClass({displayName:"LiveEditor",render:function(){var e=this,t=this.props.recipe,n=t.name,l=void 0===n?"":n,a=t.ingredients,u=void 0===a?[]:a,i=t.instructions,c=void 0===i?[]:i;t.setName,t.setIngredients,t.setInstructions;return r.default.createElement("div",{className:"live-editor"},r.default.createElement(s,{name:"name",change:function(t){e.props.recipe.name=t},value:l}),r.default.createElement(o,{name:"ingredients",change:function(t){e.props.recipe.ingredients=t},value:u.join("\n")}),r.default.createElement(o,{name:"instructions",change:function(t){e.props.recipe.instructions=t},value:c.join("\n")}))}});t.default=(0,u.observer)(i);var s=(0,u.observer)(r.default.createClass({displayName:"UserInput",propTypes:{value:r.default.PropTypes.string.isRequired,name:r.default.PropTypes.string.isRequired,change:r.default.PropTypes.func.isRequired},handleChange:function(e){this.props.change(e.target.value)},handleBlur:function(e){this.props.change(e.target.value)},handleSubmit:function(e){13===e.which&&this.props.change(e.target.value)},render:function(){return r.default.createElement("div",null,r.default.createElement("label",null,this.props.name,r.default.createElement("input",{type:"text",value:this.props.value,onChange:this.handleChange,onBlur:this.handleBlur,onKeyDown:this.handleSubmit})))}})),c=function(e){return e.split("\n")},o=(0,u.observer)(r.default.createClass({displayName:"UserTextarea",handleChange:function(e){this.set(e.target.value)},handleSubmit:function(e){13===e.which&&this.set(e.target.value)},set:function(e){this.props.change(c(e))},render:function(){return r.default.createElement("div",null,r.default.createElement("label",null,this.props.name,r.default.createElement("textarea",{value:this.props.value,onChange:this.handleChange,onBlur:this.handleBlur,onKeyDown:this.handleSubmit})))}}))},168:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=n(1),r=l(a),u=n(164),i=n(169),s=l(i),c=n(170),o=l(c),d=function(e){var t=e.recipe;return r.default.createElement("div",{className:"recipe"},r.default.createElement("h2",null,t.name),r.default.createElement("h3",null,""!==t.ytID?"https://youtu.be/"+t.ytID:null),r.default.createElement(s.default,{values:t.ingredients||[]}),r.default.createElement(o.default,{values:t.instructions||[]}))};t.default=(0,u.observer)(d)},169:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}function a(e){return u.default.createElement("div",{className:"ingredients"},u.default.createElement("h3",null,"Ingredients"),u.default.createElement("ul",null,e.values.filter(function(e){return""!==e}).map(function(e,t){return u.default.createElement("li",{key:t},e)})))}Object.defineProperty(t,"__esModule",{value:!0});var r=n(1),u=l(r),i=n(164);t.default=(0,i.observer)(a)},170:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}function a(e){return u.default.createElement("div",{className:"instructions"},u.default.createElement("h3",null,"Instructions"),u.default.createElement("ol",null,e.values.filter(function(e){return""!==e}).map(function(e,t){return u.default.createElement("li",{key:t},e)})))}Object.defineProperty(t,"__esModule",{value:!0});var r=n(1),u=l(r),i=n(164);t.default=(0,i.observer)(a)},171:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=n(1),r=l(a);t.default=r.default.createClass({displayName:"video",propTypes:{ytID:r.default.PropTypes.string.isRequired},shouldComponentUpdate:function(e,t){return e.ytID!==this.props.ytID},render:function(){var e="https://www.youtube.com/embed/"+this.props.ytID,t=""===this.props.ytID?null:r.default.createElement("iframe",{width:"560",height:"315",src:e,frameBorder:"0"});return r.default.createElement("div",{className:"yt"},t)}})},172:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var l in n)Object.prototype.hasOwnProperty.call(n,l)&&(e[l]=n[l])}return e},r=n(1),u=l(r),i=n(164),s=n(173),c=function(e){var t=e.recipes,n=t.recipes,l=t.index,r=n.map(function(t,n){return null===t?null:u.default.createElement(o,a({key:n,index:n,active:n===l,delete:function(){e.recipes.recipes[n]=null},load:function(){e.recipes.index=n}},t))});return u.default.createElement("div",{className:"recipe-menu"},"Saved Recipes:",u.default.createElement("ul",{className:"saved-recipes"},r),u.default.createElement(d,{recipes:e.recipes}))};t.default=(0,i.observer)(c);var o=(0,i.observer)(u.default.createClass({displayName:"Thumbnail",handleDelete:function(e){e.stopPropagation(),this.props.delete(this.props.index)},handleClick:function(e){this.props.load(this.props.index)},render:function(){var e=this.props,t=e.ytID,n=e.name,l=e.active,a="https://i.ytimg.com/vi/"+t+"/mqdefault.jpg",r=""===t?u.default.createElement("div",{className:"empty-thumb"},"?"):u.default.createElement("img",{src:a}),i="thumbnail";return l&&(i+=" active"),u.default.createElement("li",{className:i,onClick:this.handleClick},u.default.createElement("div",null,r),u.default.createElement("div",{className:"thumb-info"},n),u.default.createElement("div",{className:"thumb-controls"},u.default.createElement("button",{title:"delete recipe",onClick:this.handleDelete},String.fromCharCode(215))))}})),d=(0,i.observer)(u.default.createClass({displayName:"RecipeCreator",getInitialState:function(){return{value:""}},createRecipe:function(e){e.preventDefault();var t=(0,s.VideoID)(this.state.value),n=this.props.recipes,l=n.recipes,a=n.index;l.push({name:"",ytID:t,ingredients:[],instructions:[]}),a=l.length-1,this.setState({value:""})},handleChange:function(e){this.setState({value:e.target.value})},render:function(){return u.default.createElement("div",null,u.default.createElement("input",{placeholder:"YouTube URL...",value:this.state.value,onChange:this.handleChange}),u.default.createElement("button",{onClick:this.createRecipe},"Add Recipe"))}}))},173:function(e,t){"use strict";function n(e,t){localStorage.setItem(e,JSON.stringify(t))}function l(e){return JSON.parse(localStorage.getItem(e))}function a(e){var t="",n=document.createElement("a");switch(n.href=e,n.hostname){case"www.youtube.com":if(""===n.search)break;var l=n.search.slice(1).split("&");l.some(function(e){var n=e.split("=");return"v"===n[0]&&(t=n[1],!0)});break;case"youtu.be":var l=n.pathname.split("/");t=l[l.length-1]}return t}Object.defineProperty(t,"__esModule",{value:!0}),t.save=n,t.load=l,t.VideoID=a},174:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e){return r.default.createElement("header",null,r.default.createElement("h1",null,"Annotated Meals"),r.default.createElement("p",null,"Did you see a recipe on YouTube that you'd like to make, but don't want to constantly pause the video while you're cooking? Annotated Meals lets you quickly type up a recipe from a video."),r.default.createElement(i.default,null),r.default.createElement("p",null,"For a quick test, try pasting this link ",r.default.createElement("strong",null,"https://www.youtube.com/watch?v=bjmYkPkjnVo"),' into the "YouTube URL..." input below.'))};var a=n(1),r=l(a),u=n(175),i=l(u)},175:function(e,t,n){"use strict";function l(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var a=n(1),r=l(a);t.default=r.default.createClass({displayName:"howto",getInitialState:function(){return{visible:!0}},shouldComponentUpdate:function(e,t){return t.visible!==this.state.visible},toggleVisible:function(e){e.preventDefault(),this.setState({visible:!this.state.visible})},render:function(){var e=this.state.visible?"Hide Help":"Show Help",t=this.state.visible?"":"hidden";return r.default.createElement("div",{className:"help"},r.default.createElement("h3",null,"How to:"),r.default.createElement("button",{onClick:this.toggleVisible},e),r.default.createElement("ol",{className:t},r.default.createElement("li",null,"Go to the YouTube page for the recipe you would like to make and copy the url from the address bar."),r.default.createElement("li",null,'Paste the url into the "YouTube URL..." box below, then click "Add Recipe"'),r.default.createElement("li",null,"An embedded copy of the video will appear in the page."),r.default.createElement("li",null,"Type in the name of the recipe."),r.default.createElement("li",null,"Press play on the video and write down the ingredients and instructions while you watch the video, pausing to catch up whenever the chef is moving faster than you can type."),r.default.createElement("li",null,'When you\'ve written everything down, click the "Save" button.'),r.default.createElement("li",null,"If you would like a paper copy, print the recipe. Only your input will be printed, everything else will be hidden.")))}})}});
